# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API description",
        "title": "API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Выполняет вход по email и паролю, возвращает пару токенов.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Аутентификация",
                "parameters": [
                    {
                        "description": "Учётные данные",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Неверный email или пароль",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Инвалидирует переданный refresh token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Выход из текущей сессии",
                "parameters": [
                    {
                        "description": "Refresh token текущей сессии",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LogoutResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout-all": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Инвалидирует все refresh token текущего пользователя и увеличивает версию сессии.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Выход со всех устройств",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LogoutResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает данные авторизованного пользователя по JWT.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Текущий пользователь",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.GetMeResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Принимает refresh token и user_id, возвращает новую пару access + refresh.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновление токенов",
                "parameters": [
                    {
                        "description": "Refresh token и ID пользователя",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Невалидный или истёкший refresh token",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/registration": {
            "post": {
                "description": "Создаёт нового пользователя с ролью student и возвращает пару токенов.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Самостоятельная регистрация",
                "parameters": [
                    {
                        "description": "Данные регистрации",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegistrationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/auth.RegistrationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Email уже зарегистрирован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/registration/admin": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт пользователя с указанной ролью и генерирует временный пароль. Для school_admin school_id подставляется автоматически из токена.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Создание пользователя админом",
                "parameters": [
                    {
                        "description": "Данные нового пользователя",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegistrationRequestByAdmin"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/auth.RegistrationResponseByAdmin"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Email уже зарегистрирован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.GetMeResponse": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "role"
            ],
            "properties": {
                "avatar_key": {
                    "type": "string",
                    "example": "avatars/10cb44c1/photo.png"
                },
                "class_id": {
                    "type": "string",
                    "example": "1db34ac1-23f1-4b47-82e3-9e8ad16ee1fb"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-03-16T10:30:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "example@example.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "Иван"
                },
                "id": {
                    "type": "string",
                    "example": "10cb44c1-18f7-4a3e-b0bd-5d2609619d65"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "last_name": {
                    "type": "string",
                    "example": "Иванов"
                },
                "middle_name": {
                    "type": "string",
                    "example": "Иванович"
                },
                "role": {
                    "enum": [
                        "student",
                        "teacher",
                        "school_admin",
                        "super_admin"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.UserRole"
                        }
                    ],
                    "example": "student"
                },
                "school_id": {
                    "type": "string",
                    "example": "3e1b8139-1f9d-458f-ac70-41f815c8b128"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-03-16T10:30:00Z"
                }
            }
        },
        "auth.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "example@example.com"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "NFMpC9!fm;ARqoh"
                }
            }
        },
        "auth.LoginResponse": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "role"
            ],
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "class_id": {
                    "type": "string",
                    "example": "1db34ac1-23f1-4b47-82e3-9e8ad16ee1fb"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-03-16T10:30:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "example@example.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "Иван"
                },
                "id": {
                    "type": "string",
                    "example": "10cb44c1-18f7-4a3e-b0bd-5d2609619d65"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "last_name": {
                    "type": "string",
                    "example": "Иванов"
                },
                "middle_name": {
                    "type": "string",
                    "example": "Иванович"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "b7d0e8d1-47d8-4d3e-8c2b-7b1d0e8a1234"
                },
                "role": {
                    "enum": [
                        "student",
                        "teacher",
                        "school_admin",
                        "super_admin"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.UserRole"
                        }
                    ],
                    "example": "student"
                },
                "school_id": {
                    "type": "string",
                    "example": "3e1b8139-1f9d-458f-ac70-41f815c8b128"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-03-16T10:30:00Z"
                }
            }
        },
        "auth.LogoutRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "b7d0e8d1-47d8-4d3e-8c2b-7b1d0e8a1234"
                }
            }
        },
        "auth.LogoutResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "10cb44c1-18f7-4a3e-b0bd-5d2609619d65"
                }
            }
        },
        "auth.RefreshRequest": {
            "type": "object",
            "required": [
                "refresh_token",
                "user_id"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "b7d0e8d1-47d8-4d3e-8c2b-7b1d0e8a1234"
                },
                "user_id": {
                    "type": "string",
                    "example": "10cb44c1-18f7-4a3e-b0bd-5d2609619d65"
                }
            }
        },
        "auth.RefreshResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "refresh_token": {
                    "type": "string",
                    "example": "b7d0e8d1-47d8-4d3e-8c2b-7b1d0e8a1234"
                }
            }
        },
        "auth.RegistrationRequest": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "example@example.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "Иван"
                },
                "last_name": {
                    "type": "string",
                    "example": "Иванов"
                },
                "middle_name": {
                    "type": "string",
                    "example": "Иванович"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "NFMpC9!fm;ARqoh"
                }
            }
        },
...
        "auth.RegistrationResponse": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "role"
            ],
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "class_id": {
                    "type": "string",
                    "example": "1db34ac1-23f1-4b47-82e3-9e8ad16ee1fb"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-03-16T10:30:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "example@example.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "Иван"
                },
                "id": {
                    "type": "string",
                    "example": "10cb44c1-18f7-4a3e-b0bd-5d2609619d65"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "last_name": {
                    "type": "string",
                    "example": "Иванов"
                },
                "middle_name": {
                    "type": "string",
                    "example": "Иванович"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "b7d0e8d1-47d8-4d3e-8c2b-7b1d0e8a1234"
                },
                "role": {
                    "enum": [
                        "student",
                        "teacher",
                        "school_admin",
                        "super_admin"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.UserRole"
                        }
                    ],
                    "example": "student"
                },
                "school_id": {
                    "type": "string",
                    "example": "3e1b8139-1f9d-458f-ac70-41f815c8b128"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-03-16T10:30:00Z"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "message": {
                    "type": "string",
                    "example": "Некорректный формат email"
                },
                "request_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "dto.UserRole": {
            "type": "string",
            "enum": [
                "student",
                "teacher",
                "school_admin",
                "super_admin"
            ],
            "x-enum-varnames": [
                "RoleStudent",
                "RoleTeacher",
                "RoleSchoolAdmin",
                "RoleSuperAdmin"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer {access_token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}